Create helpers.py
Functions related to the backend

create save_plot_image() to create a seaborn lineplot for each table. plots correspond to apr values pulled from sqlite database and given as a pandas Dataframe.

figure is generated and saved as an .png image in static folder for use to display on the frontend